refactor: déplacement DynamicMapPopupContent dans un fichier à part
